Introduce common Message class
